requirements.txt
yt-dlp
deno
PyYAML
Pillow
requests
